@@ -25,7 +25,7 @@ public class CustomHolderDialogsActivity extends DemoDialogsActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_custom_holder_dialogs);
 
-        dialogsList = (DialogsList) findViewById(R.id.dialogsList);
+        dialogsList = findViewById(R.id.dialogsList);
         initAdapter();
     }
 
