@@ -36,10 +36,10 @@ public class CustomMediaMessagesActivity extends DemoMessagesActivity
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_custom_media_messages);
 
-        this.messagesList = (MessagesList) findViewById(R.id.messagesList);
+        this.messagesList = findViewById(R.id.messagesList);
         initAdapter();
 
-        MessageInput input = (MessageInput) findViewById(R.id.input);
+        MessageInput input = findViewById(R.id.input);
         input.setInputListener(this);
         input.setAttachmentsListener(this);
     }
@@ -60,11 +60,10 @@ public class CustomMediaMessagesActivity extends DemoMessagesActivity
 
     @Override
     public boolean hasContentFor(Message message, byte type) {
-        switch (type) {
-            case CONTENT_TYPE_VOICE:
-                return message.getVoice() != null
-                        && message.getVoice().getUrl() != null
-                        && !message.getVoice().getUrl().isEmpty();
+        if (type == CONTENT_TYPE_VOICE) {
+            return message.getVoice() != null
+                    && message.getVoice().getUrl() != null
+                    && !message.getVoice().getUrl().isEmpty();
         }
         return false;
     }
